stm32_chatbot_project/requirements.txt
transformers
torch
sentence-transformers
faiss-cpu
langchain
huggingface_hub[hf_xet]
accelerate
bitsandbytes
gradio
